% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_functions.R
\name{add_seasons}
\alias{add_seasons}
\title{Add seasons}
\usage{
add_seasons(dataset)
}
\arguments{
\item{dataset}{data frame}
}
\value{
Add a variable \strong{Season}, which has \emph{Winter}, \emph{Spring}, \emph{Summer} and \emph{Autumn}
}
\description{
Use for adding seasons, to look if in each season there is a pattern
}
