% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_functions.R
\name{combine_meteo_data}
\alias{combine_meteo_data}
\title{Combine meteo dataset}
\usage{
combine_meteo_data(first_dataset, second_dataset)
}
\arguments{
\item{first_dataset}{data.frame}

\item{second_dataset}{data.frame}
}
\description{
Combine meteo dataset
}
