% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_functions.R
\name{first_graph_analysis}
\alias{first_graph_analysis}
\title{First ggplot}
\usage{
first_graph_analysis(dataset)
}
\arguments{
\item{dataset}{data frame}
}
\value{
a ggplot of the \code{dataset}
}
\description{
A first glance on the data to have an idea of the evolution of the temperature
}
\details{
Expected the data is the combination of two weather data frame
}
