% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_functions.R
\name{second_graph_analysis}
\alias{second_graph_analysis}
\title{Second plot, ggpairs}
\usage{
second_graph_analysis(dataset)
}
\arguments{
\item{dataset}{data.frame}
}
\value{
a ggppairs of \code{dataset}
}
\description{
a more general view on other variables
}
\details{
Expected the data is the combination of two weather data frame
}
