% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_functions.R
\name{third_graph_analysis}
\alias{third_graph_analysis}
\title{Autoplot, for validation}
\usage{
third_graph_analysis(linear_model)
}
\arguments{
\item{linear_model}{lm}
}
\value{
the autoplot of \code{linear_model}
}
\description{
a validation of \code{linear model}
}
