% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_functions.R
\name{thrid_graph_analysis}
\alias{thrid_graph_analysis}
\title{Autoplot, for validation}
\usage{
thrid_graph_analysis(linear_model)
}
\arguments{
\item{linear_model}{lm}
}
\description{
Autoplot, for validation
}
