% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_functions.R
\name{best_linear_model_in_this_analysis}
\alias{best_linear_model_in_this_analysis}
\title{Find suitable linear model}
\usage{
best_linear_model_in_this_analysis(dataset)
}
\arguments{
\item{dataset}{data.frame}
}
\description{
Find suitable linear model
}
